Tracker_simulation/requirements.txt
kafka-python==2.2.4
numpy==2.2.5
matplotlib
opencv-python